refactor: extract mutation campaign catalog
